requirements.txt
playwright==1.44.0
pyTelegramBotAPI==4.21.0
schedule==1.2.2
python-dotenv==1.0.1
